Default prefer_commandline to true.
My first instinct was that it should be true, but I wasn't sure and
MooX::Options defaults to false, so I went with false. mst convinced me
that true makes a lot more sense -- it works better for specializing
scripts, and it means that "with_config_from_file" support is easily
replaced by a small config loader in the script.